// src/types/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Long: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export enum Enum_Fiadocument_Type {
  Info = 'info',
  InvestigationNoFurtherAction = 'investigation_no_further_action',
  InvestigationPenaltyGridDrop = 'investigation_penalty_grid_drop',
  InvestigationPenaltyTime = 'investigation_penalty_time',
  InvestigationRightOfReviewPenaltyGridDrop = 'investigation_right_of_review_penalty_grid_drop',
  InvestigationRightOfReviewPenaltyNoFurtherAction = 'investigation_right_of_review_penalty_no_further_action',
  InvestigationRightOfReviewPenaltyTime = 'investigation_right_of_review_penalty_time'
}

export enum Enum_Schedulerscheduler_Type {
  Archive = 'archive',
  Publish = 'publish'
}

export enum Enum_Session_Type {
  Practice = 'practice',
  Qualifying = 'qualifying',
  Race = 'race',
  Session = 'session'
}
// ...
export type FiaDocumentRelationResponseCollection = {
  __typename?: 'FiaDocumentRelationResponseCollection';
  data: Array<FiaDocumentEntity>;
};
// ...
export type GrandPrixEntityResponse = {
  __typename?: 'GrandPrixEntityResponse';
  data?: Maybe<GrandPrixEntity>;
};
// ...
export type Session = {
  __typename?: 'Session';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  endTime?: Maybe<Scalars['DateTime']['output']>;
  fia_documents?: Maybe<FiaDocumentRelationResponseCollection>;
  grands_prix?: Maybe<GrandPrixEntityResponse>;
  name?: Maybe<Scalars['String']['output']>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  startTime?: Maybe<Scalars['DateTime']['output']>;
  type?: Maybe<Enum_Session_Type>;
  uid?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};